ios: fix thread access warning
The subviews property needs to be called for every rendered frame, which
is called from the render thread (not the UI thread) thus we cannot
access UI stuff such as the subviews of a given UIView.

However, there is no need to do it since we already know the subview.
Storing the extra reference should not be a problem since we are already
involved in a cycle which is broken when the view is destroyed anyway.